requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
openai>=1.40.0
pydantic>=2.0.0
json-repair>=0.30.0
python-dotenv>=1.0.0
